refactor(weather): Update repository imports in view model layer
- Modified import paths in city_search_view_model.dart, city_weather_notifier.dart,
and weather_view_model.dart to reference repositories under features/weather/.
- Ensures consistency with the new feature-based directory structure.
